% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef}
\alias{coef}
\alias{coef.sfacross}
\alias{coef.summary.sfacross}
\alias{coef.lcmcross}
\alias{coef.summary.lcmcross}
\alias{coef.sfaselectioncross}
\alias{coef.summary.sfaselectioncross}
\alias{coef.zisfcross}
\alias{coef.summary.zisfcross}
\alias{coef.cnsfcross}
\alias{coef.summary.cnsfcross}
\alias{coef.misfcross}
\alias{coef.summary.misfcross}
\title{Extract coefficients of stochastic frontier models}
\usage{
\method{coef}{sfacross}(object, extraPar = FALSE, ...)

\method{coef}{summary.sfacross}(object, ...)

\method{coef}{lcmcross}(object, extraPar = FALSE, ...)

\method{coef}{summary.lcmcross}(object, ...)

\method{coef}{sfaselectioncross}(object, extraPar = FALSE, ...)

\method{coef}{summary.sfaselectioncross}(object, ...)

\method{coef}{zisfcross}(object, extraPar = FALSE, ...)

\method{coef}{summary.zisfcross}(object, ...)

\method{coef}{cnsfcross}(object, extraPar = FALSE, ...)

\method{coef}{summary.cnsfcross}(object, ...)

\method{coef}{misfcross}(object, extraPar = FALSE, ...)

\method{coef}{summary.misfcross}(object, ...)
}
\arguments{
\item{object}{A stochastic frontier model returned
by \code{\link{cnsfcross}}, \code{\link{lcmcross}}, \code{\link{misfcross}},
\code{\link{sfacross}}, \code{\link{sfaselectioncross}} or
\code{\link{zisfcross}}, or an object of class \code{'summary.cnsfcross'},
\code{'summary.lcmcross'}, \code{'summary.misfcross'},
\code{'summary.sfacross'}, \code{'summary.sfaselectioncross'} or
\code{'summary.zisfcross'} .}

\item{extraPar}{Logical (default = \code{FALSE}). If \code{TRUE}, additional
parameters are returned:

\code{sigmaSq} = \code{sigmauSq} + \code{sigmavSq}

\code{lambdaSq} = \code{sigmauSq}/\code{sigmavSq}

\code{sigmauSq} = \eqn{\exp{(Wu)}} = \eqn{\exp{(\delta Z_u)}}

\code{sigmavSq} = \eqn{\exp{(Wv)}} = \eqn{\exp{(\phi Z_v)}}

\code{sigma} = \code{sigmaSq}^0.5

\code{lambda} = \code{lambdaSq}^0.5

\code{sigmau} = \code{sigmauSq}^0.5

\code{sigmav} = \code{sigmavSq}^0.5

\code{gamma} = \code{sigmauSq}/(\code{sigmauSq} + \code{sigmavSq})}

\item{...}{Currently ignored.}
}
\value{
For objects of class \code{'summary.cnsfcross'},
\code{'summary.lcmcross'},  \code{'summary.sfacross'},
\code{'summary.sfaselectioncross'} or \code{'summary.zisfcross'},
\code{\link{coef}} returns a matrix with four columns. Namely, the estimated
coefficients, their standard errors, z-values, and (asymptotic) P-values.

For objects of class \code{'cnsfcross'}, \code{'lcmcross'},
\code{'sfacross'}, \code{'sfaselectioncross'} or \code{'zisfcross'},
\code{\link{coef}} returns a numeric vector of the estimated coefficients.
If \code{extraPar = TRUE}, additional parameters, detailed in the section
\sQuote{Arguments}, are also returned. In the case of object of class
\code{'lcmcross'}, each additional parameter terminates with \code{'#'} that
represents the class number.
}
\description{
From an object of class \code{'summary.cnsfcross'}, \code{'summary.lcmcross'},
\code{'summary.misfcross'}, \code{'summary.sfacross'},
\code{'summary.sfaselectioncross'} or \code{'summary.zisfcross'},
\code{\link{coef}} extracts the coefficients,
their standard errors, z-values, and (asymptotic) P-values.

From on object of class \code{'cnsfcross'}, \code{'lcmcross'},
\code{'misfcross'}, \code{'sfacross'}, \code{'sfaselectioncross'} or
\code{'zisfcross'}, it extracts only the estimated coefficients.
}
\examples{

## Using data on fossil fuel fired steam electric power generation plants in the U.S.
# Translog SFA (cost function) truncated normal with scaling property
tl_u_ts <- sfacross(formula = log(tc/wf) ~ log(y) + I(1/2 * (log(y))^2) +
log(wl/wf) + log(wk/wf) + I(1/2 * (log(wl/wf))^2) + I(1/2 * (log(wk/wf))^2) +
I(log(wl/wf) * log(wk/wf)) + I(log(y) * log(wl/wf)) + I(log(y) * log(wk/wf)),
udist = 'tnormal', muhet = ~ regu, uhet = ~ regu, data = utility, S = -1,
scaling = TRUE, method = 'mla')
coef(tl_u_ts, extraPar = TRUE)
coef(summary(tl_u_ts))

}
\seealso{
\code{\link{cnsfcross}}, for the contaminated noise stochastic
frontier analysis model fitting function.

\code{\link{lcmcross}}, for the latent class stochastic frontier analysis
model fitting function.

\code{\link{misfcross}}, for the multi-modal inefficiency stochastic frontier
analysis model fitting function.

\code{\link{sfacross}}, for the stochastic frontier analysis model
fitting function.

\code{\link{sfaselectioncross}} for sample selection in stochastic frontier
model fitting function.

\code{\link{zisfcross}} for zero inefficiency in stochastic frontier model
fitting function.
}
\keyword{coefficients}
\keyword{methods}
