% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{sfaR-deprecated}
\alias{sfaR-deprecated}
\alias{lcmcross}
\alias{print.lcmcross}
\alias{bread.lcmcross}
\alias{estfun.lcmcross}
\alias{coef.lcmcross}
\alias{coef.summary.lcmcross}
\alias{fitted.lcmcross}
\alias{ic.lcmcross}
\alias{logLik.lcmcross}
\alias{marginal.lcmcross}
\alias{nobs.lcmcross}
\alias{residuals.lcmcross}
\alias{summary.lcmcross}
\alias{print.summary.lcmcross}
\alias{efficiencies.lcmcross}
\alias{vcov.lcmcross}
\title{Deprecated functions of sfaR}
\usage{
lcmcross(
  formula,
  uhet,
  vhet,
  thet,
  logDepVar = TRUE,
  data,
  subset,
  weights,
  wscale = TRUE,
  S = 1L,
  udist = "hnormal",
  start = NULL,
  whichStart = 2L,
  initAlg = "nm",
  initIter = 100,
  lcmClasses = 2,
  method = "bfgs",
  hessianType = 1,
  itermax = 2000L,
  printInfo = FALSE,
  tol = 1e-12,
  gradtol = 1e-06,
  stepmax = 0.1,
  qac = "marquardt"
)

\method{print}{lcmcross}(x, ...)

\method{bread}{lcmcross}(x, ...)

\method{estfun}{lcmcross}(x, ...)

\method{coef}{lcmcross}(object, extraPar = FALSE, ...)

\method{coef}{summary.lcmcross}(object, ...)

\method{fitted}{lcmcross}(object, ...)

\method{ic}{lcmcross}(object, IC = "AIC", ...)

\method{logLik}{lcmcross}(object, individual = FALSE, ...)

\method{marginal}{lcmcross}(object, newData = NULL, ...)

\method{nobs}{lcmcross}(object, ...)

\method{residuals}{lcmcross}(object, ...)

\method{summary}{lcmcross}(object, grad = FALSE, ci = FALSE, ...)

\method{print}{summary.lcmcross}(x, digits = max(3, getOption("digits") - 2), ...)

\method{efficiencies}{lcmcross}(object, level = 0.95, newData = NULL, ...)

\method{vcov}{lcmcross}(object, ...)
}
\arguments{
\item{formula}{A symbolic description of the model to be estimated based on
the generic function \code{formula} (see section \sQuote{Details}).}

\item{uhet}{A one-part formula to account for heteroscedasticity in the
one-sided error variance (see section \sQuote{Details}).}

\item{vhet}{A one-part formula to account for heteroscedasticity in the
two-sided error variance (see section \sQuote{Details}).}

\item{thet}{A one-part formula to account for technological heterogeneity in
the construction of the classes.}

\item{logDepVar}{Logical. Informs whether the dependent variable is logged
(\code{TRUE}) or not (\code{FALSE}). Default = \code{TRUE}.}

\item{data}{The data frame containing the data.}

\item{subset}{An optional vector specifying a subset of observations to be
used in the optimization process.}

\item{weights}{An optional vector of weights to be used for weighted
log-likelihood. Should be \code{NULL} or numeric vector with positive values.
When \code{NULL}, a numeric vector of 1 is used.}

\item{wscale}{Logical. When \code{weights} is not \code{NULL}, a scaling
transformation is used such that the \code{weights} sums to the sample
size. Default \code{TRUE}. When \code{FALSE} no scaling is used.}

\item{S}{If \code{S = 1} (default), a production (profit) frontier is
estimated: \eqn{\epsilon_i = v_i-u_i}. If \code{S = -1}, a cost frontier is
estimated: \eqn{\epsilon_i = v_i+u_i}.}

\item{udist}{Character string. Distribution specification for the one-sided
error term. Only the half normal distribution \code{'hnormal'} (Aigner
\emph{et al.}, 1977, Meeusen and Vandenbroeck, 1977) is currently
implemented.}

\item{start}{Numeric vector. Optional starting values for the maximum
likelihood (ML) estimation.}

\item{whichStart}{Integer. If \code{'whichStart = 1'}, the starting values
are obtained from the method of moments. When \code{'whichStart = 2'}
(Default), the model is initialized by solving the homoscedastic pooled
cross section SFA model. \code{'whichStart = 1'} can be fast.}

\item{initAlg}{Character string specifying the algorithm used for
initialization and obtain the starting values (when \code{'whichStart = 2'}).
Only \pkg{maxLik} package algorithms are available:
\itemize{ \item \code{'bfgs'}, for Broyden-Fletcher-Goldfarb-Shanno
(see \code{\link[maxLik:maxBFGS]{maxBFGS}})
\item \code{'bhhh'}, for Berndt-Hall-Hall-Hausman
(see \code{\link[maxLik:maxBHHH]{maxBHHH}})
\item \code{'nr'}, for Newton-Raphson (see \code{\link[maxLik:maxNR]{maxNR}})
\item \code{'nm'}, for Nelder-Mead - Default -
(see \code{\link[maxLik:maxNM]{maxNM}})
\item \code{'cg'}, for Conjugate Gradient
(see \code{\link[maxLik:maxCG]{maxCG}}) \item \code{'sann'}, for Simulated
Annealing (see \code{\link[maxLik:maxSANN]{maxSANN}})
}}

\item{initIter}{Maximum number of iterations for initialization algorithm.
Default \code{100}.}

\item{lcmClasses}{Number of classes to be estimated (default = \code{2}). A
maximum of five classes can be estimated.}

\item{method}{Optimization algorithm used for the estimation.  Default =
\code{'bfgs'}. 11 algorithms are available: \itemize{ \item \code{'bfgs'},
for Broyden-Fletcher-Goldfarb-Shanno (see
\code{\link[maxLik:maxBFGS]{maxBFGS}}) \item \code{'bhhh'}, for
Berndt-Hall-Hall-Hausman (see \code{\link[maxLik:maxBHHH]{maxBHHH}}) \item
\code{'nr'}, for Newton-Raphson (see \code{\link[maxLik:maxNR]{maxNR}})
\item \code{'nm'}, for Nelder-Mead (see \code{\link[maxLik:maxNM]{maxNM}})
\item \code{'cg'}, for Conjugate Gradient
(see \code{\link[maxLik:maxCG]{maxCG}}) \item \code{'sann'}, for Simulated
Annealing (see \code{\link[maxLik:maxSANN]{maxSANN}})
\item \code{'ucminf'}, for a quasi-Newton type optimization with BFGS updating of
the inverse Hessian and soft line search with a trust region type monitoring
of the input to the line search algorithm
(see \code{\link[ucminf:ucminf]{ucminf}})
\item \code{'mla'}, for general-purpose optimization based on
Marquardt-Levenberg algorithm (see \code{\link[marqLevAlg:mla]{mla}})
\item \code{'sr1'}, for Symmetric Rank 1 (see
\code{\link[trustOptim:trust.optim]{trust.optim}}) \item \code{'sparse'},
for trust regions and sparse Hessian
(see \code{\link[trustOptim:trust.optim]{trust.optim}}) \item
\code{'nlminb'}, for optimization using PORT routines (see
\code{\link[stats:nlminb]{nlminb}})}}

\item{hessianType}{Integer. If \code{1} (default), analytic Hessian is
returned. If \code{2}, bhhh Hessian is estimated (\eqn{g'g}).}

\item{itermax}{Maximum number of iterations allowed for optimization.
Default = \code{2000}.}

\item{printInfo}{Logical. Print information during optimization. Default =
\code{FALSE}.}

\item{tol}{Numeric. Convergence tolerance. Default = \code{1e-12}.}

\item{gradtol}{Numeric. Convergence tolerance for gradient. Default =
\code{1e-06}.}

\item{stepmax}{Numeric. Step max for \code{ucminf} algorithm. Default =
\code{0.1}.}

\item{qac}{Character. Quadratic Approximation Correction for \code{'bhhh'}
and \code{'nr'} algorithms. If \code{'qac = stephalving'}, the step length
is decreased but the direction is kept. If \code{'qac = marquardt'}
(default), the step length is decreased while also moving closer to the pure
gradient direction. See \code{\link[maxLik:maxBHHH]{maxBHHH}} and
\code{\link[maxLik:maxNR]{maxNR}}.}

\item{x}{an object of class lcmcross (returned by the function
\code{\link{lcmcross}}).}

\item{...}{additional arguments of frontier are passed to lcmcross;
additional arguments of the print, bread, estfun, nobs methods are currently
ignored.}

\item{object}{an object of class lcmcross (returned by the function
\code{\link{lcmcross}}).}

\item{extraPar}{Logical (default = \code{FALSE}). If \code{TRUE}, additional
parameters are returned (see \code{\link{coef}} or \code{\link{vcov}}).}

\item{IC}{Character string. Information criterion measure. Three criteria
are available: \itemize{ \item \code{'AIC'} for Akaike information criterion
(default) \item \code{'BIC'} for Bayesian information criterion \item
\code{'HQIC'} for Hannan-Quinn information criterion }.}

\item{individual}{Logical. If \code{FALSE} (default), the sum of all
observations' log-likelihood values is returned. If \code{TRUE}, a vector of
each observation's log-likelihood value is returned.}

\item{newData}{Optional data frame that is used to calculate the efficiency
estimates. If NULL (the default), the efficiency estimates are calculated
for the observations that were used in the estimation.}

\item{grad}{Logical. Default = \code{FALSE}. If \code{TRUE}, the gradient
for the maximum likelihood (ML) estimates of the different parameters is
returned.}

\item{ci}{Logical. Default = \code{FALSE}. If \code{TRUE}, the 95\%
confidence interval for the different parameters (OLS or/and ML estimates) is
returned.}

\item{digits}{Numeric. Number of digits displayed in values.}

\item{level}{A number between between 0 and 0.9999 used for the computation
of (in-)efficiency confidence intervals (defaut = \code{0.95}). Not used in the
case of \code{lcmcross}.}
}
\description{
These functions are provided for compatibility with older versions of
\sQuote{sfaR} only, and could be defunct at a future release.
}
\details{
The following functions are deprecated and could be removed from \pkg{sfaR}
in a near future. Use the replacement indicated below:
\itemize{
\item{lcmcross: \code{\link{sfalcmcross}}}
\item{bread.lcmcross: \code{\link{bread.sfalcmcross}}}
\item{coef.lcmcross: \code{\link{coef.sfalcmcross}}}
\item{coef.summary.lcmcross: \code{\link{coef.summary.sfalcmcross}}}
\item{efficiencies.lcmcross: \code{\link{efficiencies.sfalcmcross}}}
\item{estfun.lcmcross: \code{\link{estfun.sfalcmcross}}}
\item{fitted.lcmcross: \code{\link{fitted.sfalcmcross}}}
\item{ic.lcmcross: \code{\link{ic.sfalcmcross}}}
\item{logLik.lcmcross: \code{\link{logLik.sfalcmcross}}}
\item{marginal.lcmcross: \code{\link{marginal.sfalcmcross}}}
\item{nobs.lcmcross: \code{\link{nobs.sfalcmcross}}}
\item{print.lcmcross: \code{\link{print.sfalcmcross}}}
\item{print.summary.lcmcross: \code{\link{print.summary.sfalcmcross}}}
\item{residuals.lcmcross: \code{\link{residuals.sfalcmcross}}}
\item{summary.lcmcross: \code{\link{summary.sfalcmcross}}}
\item{vcov.lcmcross: \code{\link{vcov.sfalcmcross}}}
}
}
