% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.R
\name{fitted}
\alias{fitted}
\alias{fitted.sfacross}
\alias{fitted.lcmcross}
\alias{fitted.sfaselectioncross}
\alias{fitted.zisfcross}
\alias{fitted.cnsfcross}
\alias{fitted.misfcross}
\title{Extract fitted values of stochastic frontier models}
\usage{
\method{fitted}{sfacross}(object, ...)

\method{fitted}{lcmcross}(object, ...)

\method{fitted}{sfaselectioncross}(object, ...)

\method{fitted}{zisfcross}(object, ...)

\method{fitted}{cnsfcross}(object, ...)

\method{fitted}{misfcross}(object, ...)
}
\arguments{
\item{object}{A stochastic frontier model returned
by \code{\link{cnsfcross}}, \code{\link{lcmcross}}, \code{\link{misfcross}},
\code{\link{sfacross}}, \code{\link{sfaselectioncross}} or
\code{\link{zisfcross}}.}

\item{...}{Currently ignored.}
}
\value{
In the case of an object of class \code{'lcmcross'}, a data frame
containing the fitted values for each class is returned where each variable
terminates with \code{'_c#'}, \code{'#'} being the class number.

In the case of an object of class \code{'cnsfcross'}, \code{'misfcross'},
\code{'sfacross'}, \code{'sfaselectioncross'} or \code{'zisfcross'},
a vector of fitted values is returned.
}
\description{
\code{\link{fitted}} returns the fitted frontier values from classic or
latent class stochastic frontier models estimated with
\code{\link{cnsfcross}}, \code{\link{lcmcross}}, \code{\link{misfcross}},
\code{\link{sfacross}}, \code{\link{sfaselectioncross}} or
\code{\link{zisfcross}}.
}
\note{
The fitted values are ordered in the same way as the corresponding
observations in the dataset used for the estimation.
}
\examples{

## Using data on eighty-two countries production (DGP)
# LCM Cobb Douglas (production function) half normal distribution
cb_2c_h <- lcmcross(formula = ly ~ lk + ll + yr, udist = 'hnormal', 
data = worldprod)
fit.cb_2c_h <- fitted(cb_2c_h)
head(fit.cb_2c_h)

}
\seealso{
\code{\link{cnsfcross}}, for the contaminated noise stochastic
frontier analysis model fitting function.

\code{\link{lcmcross}}, for the latent class stochastic frontier analysis
model fitting function.

\code{\link{misfcross}}, for the multi-modal inefficiency stochastic frontier
analysis model fitting function.

\code{\link{sfacross}}, for the stochastic frontier analysis model
fitting function.

\code{\link{sfaselectioncross}} for sample selection in stochastic frontier
model fitting function.

\code{\link{zisfcross}} for zero inefficiency in stochastic frontier model
fitting function.
}
\keyword{fitted}
\keyword{methods}
