% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.R
\name{fitted}
\alias{fitted}
\alias{fitted.sfacross}
\alias{fitted.lcmcross}
\alias{fitted.selectioncross}
\title{Extract fitted values of stochastic frontier models}
\usage{
\method{fitted}{sfacross}(object, ...)

\method{fitted}{lcmcross}(object, ...)

\method{fitted}{selectioncross}(object, ...)
}
\arguments{
\item{object}{A stochastic frontier model returned
by \code{\link{sfacross}}, \code{\link{lcmcross}} or \code{\link{selectioncross}}.}

\item{...}{Currently ignored.}
}
\value{
In the case of an object of class \code{'sfacross'} or \code{'selectioncross'}, a vector of
fitted values is returned.

In the case of an object of class \code{'lcmcross'}, a data frame containing
the fitted values for each class is returned where each variable terminates
with \code{'_c#'}, \code{'#'} being the class number.
}
\description{
\code{\link{fitted}} returns the fitted frontier values from classic or
latent class stochastic frontier models estimated with
\code{\link{sfacross}}, \code{\link{lcmcross}} or \code{\link{selectioncross}}.
}
\note{
The fitted values are ordered in the same way as the corresponding
observations in the dataset used for the estimation.
}
\examples{

## Using data on eighty-two countries production (DGP)
# LCM Cobb Douglas (production function) half normal distribution
cb_2c_h <- lcmcross(formula = ly ~ lk + ll + yr, udist = 'hnormal', data = worldprod)
  fit.cb_2c_h <- fitted(cb_2c_h)
  head(fit.cb_2c_h)

}
\seealso{
\code{\link{sfacross}}, for the stochastic frontier analysis model
fitting function.

\code{\link{lcmcross}}, for the latent class stochastic frontier analysis
model fitting function.

\code{\link{selectioncross}} for sample selection in stochastic frontier model
fitting function.
}
\author{
K Hervé Dakpo, Yann Desjeux, and Laure Latruffe
}
\keyword{fitted}
\keyword{methods}
