% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic.R
\name{ic}
\alias{ic}
\alias{ic.sfacross}
\alias{ic.lcmcross}
\alias{ic.sfaselectioncross}
\alias{ic.zisfcross}
\alias{ic.cnsfcross}
\title{Extract information criteria of stochastic frontier models}
\usage{
\method{ic}{sfacross}(object, IC = "AIC", ...)

\method{ic}{lcmcross}(object, IC = "AIC", ...)

\method{ic}{sfaselectioncross}(object, IC = "AIC", ...)

\method{ic}{zisfcross}(object, IC = "AIC", ...)

\method{ic}{cnsfcross}(object, IC = "AIC", ...)
}
\arguments{
\item{object}{A stochastic frontier model returned
by \code{\link{cnsfcross}}, \code{\link{lcmcross}}, \code{\link{sfacross}},
\code{\link{sfaselectioncross}} or \code{\link{zisfcross}}.}

\item{IC}{Character string. Information criterion measure. Three criteria
are available: \itemize{ \item \code{'AIC'} for Akaike information criterion
(default) \item \code{'BIC'} for Bayesian information criterion \item
\code{'HQIC'} for Hannan-Quinn information criterion }}

\item{...}{Currently ignored.}
}
\value{
\code{\link{ic}} returns the value of the information criterion
(AIC, BIC or HQIC) of the maximum likelihood coefficients.
}
\description{
\code{\link{ic}} returns information criterion from classic or latent class
stochastic frontier models estimated with \code{\link{cnsfcross}},
\code{\link{lcmcross}}, \code{\link{sfacross}},
\code{\link{sfaselectioncross}} or \code{\link{zisfcross}}.
}
\details{
The different information criteria are computed as follows: \itemize{ \item
AIC: \eqn{-2 \log{LL} + 2 * K} \item BIC: \eqn{-2 \log{LL} + \log{N} * K}
\item HQIC: \eqn{-2 \log{LL} + 2 \log{\left[\log{N}\right]} * K} } where
\eqn{LL} is the maximum likelihood value, \eqn{K} the number of parameters
estimated and \eqn{N} the number of observations.
}
\examples{

## Using data on Swiss railway
# LCM (cost function) half normal distribution
cb_2c_u <- lcmcross(formula = LNCT ~ LNQ2 + LNQ3 + LNNET + LNPK + LNPL,
udist = 'hnormal', uhet = ~ 1, data = swissrailways, S = -1, method='ucminf')
ic(cb_2c_u)
ic(cb_2c_u, IC = 'BIC')
ic(cb_2c_u, IC = 'HQIC')

}
\seealso{
\code{\link{cnsfcross}}, for the contaminated noise stochastic
frontier analysis model fitting function.

\code{\link{lcmcross}}, for the latent class stochastic frontier analysis
model fitting function.

\code{\link{sfacross}}, for the stochastic frontier analysis model
fitting function.

\code{\link{sfaselectioncross}} for sample selection in stochastic frontier
model fitting function.

\code{\link{zisfcross}} for zero inefficiency in stochastic frontier model
fitting function.
}
\keyword{AIC}
\keyword{BIC}
\keyword{HQIC}
\keyword{methods}
