% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfaR-package.R
\docType{package}
\name{sfaR-package}
\alias{sfaR-package}
\alias{sfaR}
\title{sfaR: A package for estimating stochastic frontier models}
\description{
The \pkg{sfaR} package provides a set of tools (maximum likelihood - ML and
maximum simulated likelihood - MSL) for various specifications of stochastic
frontier analysis (SFA).
}
\details{
Four categories of important functions are available: \code{\link{sfacross}},
\code{\link{lcmcross}}, \code{\link{selectioncross}}, and \code{\link{zisfcross}},
which estimate different types of frontiers and
offer eleven alternative optimization algorithms (i.e. "bfgs", "bhhh", "nr",
"nm", "cg", "sann", "ucminf", "mla", "sr1", "sparse", "nlminb").
}
\section{lcmcross}{
 \code{\link{lcmcross}} estimates latent class stochastic
frontier models (LCM), which accounts for technological heterogeneity by
splitting the observations into a maximum number of five classes. The
classification operates based on a logit functional form that can be
specified using some covariates (namely, the separating variables allowing
the separation of observations in several classes). Only the half normal
distribution is available for the one-sided error term. Heteroscedasticity
in both error terms is possible. The choice of the number of classes can be
guided by several information criteria (i.e. AIC, BIC or HQIC).
}

\section{sfacross}{
 \code{\link{sfacross}} estimates the basic frontier for cross-sectional data
and allows for ten different distributions for the one-sided error term.
These distributions include the exponential, the Gamma, the generalized
exponential, the half normal, the lognormal, the truncated normal, the
truncated skewed Laplace, the Rayleigh, the uniform and the Weibull distributions.
In the case of the Gamma, lognormal and Weibull distributions, maximum simulated
likelihood (MSL) is used with the possibility of four specific distributions to
construct the draws: Halton, Generalized Halton, Sobol and uniform.
Heteroscedasticity in both error terms can be implemented, in addition to
heterogeneity in the truncated mean parameter in the case of the truncated
normal and lognormal distributions. In addition, in the case of the truncated normal
distribution, the scaling property can be estimated.
}

\section{selectioncross}{
 \code{\link{selectioncross}} estimates the frontier for cross-sectional data
in the presence of sample selection. The model solves the selection bias due to
the correlation between the two-sided errors terms in both the selection and the
frontier equations. The likelihood can be estimated using five different
possibilities: Gauss-Kronrod quadrature, adaptive integration over hypercubes
(hcubature and pcubature), Gauss-Hermite quadrature, and
maximum simulated likelihood. Only the half normal
distribution is available for the one-sided error term. Heteroscedasticity
in both error terms is possible.
}

\section{zisfcross}{
 \code{\link{zisfcross}} relaxes the assumption that all firms
are inefficient. The model splits the observations into two groups:
one efficient and the other inefficient. The probability of belonging to one of
these classes is based on a logit functional form that can be
specified using some covariates. As in the case of \code{\link{sfacross}}
ten different distributions for the one-sided error term are implemented.
These distributions include the exponential, the Gamma, the generalized
exponential, the half normal, the lognormal, the truncated normal, the
truncated skewed Laplace, the Rayleigh, the uniform and the Weibull distributions.
In the case of the Gamma, lognormal and Weibull distributions, maximum simulated
likelihood (MSL) is used with the possibility of four specific distributions to
construct the draws: Halton, Generalized Halton, Sobol and uniform.
Heteroscedasticity in both error terms can be implemented, in addition to
heterogeneity in the truncated mean parameter in the case of the truncated
normal and lognormal distributions.
}

\section{Bugreport}{
 Any bug or suggestion can be reported using the \code{sfaR}
tracker facilities at: \url{https://github.com/hdakpo/sfaR/issues}
}

\author{
K Hervé Dakpo, Yann Desjeux, and Laure Latruffe
}
