% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.R
\name{vcov}
\alias{vcov}
\alias{vcov.sfacross}
\alias{vcov.lcmcross}
\title{Variance-covariance matrix for frontier objects}
\usage{
\method{vcov}{sfacross}(object, extraPar = FALSE, ...)

\method{vcov}{lcmcross}(object, extraPar = FALSE, ...)
}
\arguments{
\item{object}{A (latent class) stochastic frontier model returned by
\code{\link{sfacross}} or \code{\link{lcmcross}}}

\item{extraPar}{Logical. Default = \code{FALSE}. If \code{TRUE}, and only in
the case of object of class \code{"sfacross"}, variances and covariances of
additional parameters are returned:

\code{sigmaSq} = \code{sigmauSq} + \code{sigmavSq},

\code{lambdaSq} = \code{sigmauSq}/\code{sigmavSq},

\code{sigmauSq} = \eqn{\exp{(Wu)}},

\code{sigmavSq} = \eqn{\exp{(Wv)}},

\code{sigma} = \code{sigmaSq}^0.5,

\code{lambda} = \code{lambdaSq}^0.5,

\code{sigmau} = \code{sigmauSq}^0.5,

\code{sigmav} = \code{sigmavSq}^0.5,

\code{gamma} = \code{sigmauSq}/(\code{sigmauSq} + \code{sigmavSq})}

\item{...}{Currently ignored}
}
\value{
\code{vcov} returns the covariance matrix of the maximum likelihood
coefficients.
}
\description{
Extract the variance-covariance matrix of the maximum likelihood
coefficients of frontier objects (\code{\link{sfacross}} or
\code{\link{lcmcross}}).
}
\details{
the variance-covariance matrix is obtained by the inversion of the
negative hessian matrix. Depending on the distribution and the
\code{"hessianType"} option the analytical/numeric hessian or the bhhh
hessian or the robust hessian matrix is evaluated. When \code{"extraPar =
  TRUE"} the variance-covariance of the additional parameters (in the case of
\code{\link{sfacross}}) is obtained using the delta method.
}
\examples{

## Data on Spanish dairy farms

# Cobb Douglas (production function) half normal, exponential and Rayleigh
# distributions

cb_s_h <- sfacross(
  formula = YIT ~ X1 + X2 + X3 + X4, udist = "hnormal", data
  = dairyspain, S = 1, method = "bfgs"
)

vcov(cb_s_h)

cb_s_e <- sfacross(
  formula = YIT ~ X1 + X2 + X3 + X4, udist = "exponential",
  data = dairyspain, S = 1, method = "mla"
)

vcov(cb_s_e)

cb_s_r <- sfacross(
  formula = YIT ~ X1 + X2 + X3 + X4, udist = "rayleigh",
  data = dairyspain, S = 1, method = "nlminb"
)

vcov(cb_s_r)
}
\seealso{
\code{\link{sfacross}}, for the stochastic frontier analysis model fitting
function.

\code{\link{lcmcross}}, for the latent class stochastic frontier analysis
model fitting function.
}
\author{
K Hervé Dakpo, Yann Desjeux and Laure Latruffe
}
\keyword{methods}
\keyword{vcov}
